moving model configs into separate files
